+++ changelog.md
@@ -5,6 +5,7 @@
 * Add the option `envs.<env>.localOverrides` to control whether local packages are included in an env's GHC overrides,
   independently of whether they're exposed.
 * Add an env, `hix-build-tools`, to configure packages used internally; like `cabal-install` and `hpack`.
+* Add the option `envs.<env>.inheritOverrides` to control whether to include overrides from dependency flakes.
 
 # 0.8.0